DreamStudio_BE/pyproject.toml
[project]
name = "dreamstudio-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Jae Kim",email = "rlawogus9643@gmail.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (==0.135.1)",
    "uvicorn[standard] (==0.30.6)",
    "sqlalchemy (==2.0.48)",
    "psycopg[binary] (==3.3.3)",
    "passlib (==1.7.4)",
    "argon2-cffi (==25.1.0)",
    "python-jose[cryptography] (==3.3.0)",
    "requests (==2.32.5)",
    "google-auth (==2.49.0)",
    "celery (==5.6.2)",
    "redis (==7.3.0)",
    "openai (==2.26.0)",
    "boto3 (==1.42.63)",
    "stripe (==14.4.1)",
    "python-dotenv (==1.0.1)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false